cli: Move createZabbixClient to utils
